fix(ci): use canonical model ids for anna, refresh renamed agent slugs, split oversized files
The rename-agents commit (c674f945) broke two required checks on main:

- ci (agentcatalog): anna's manifest referenced provider-namespaced model
  ids (`xiaomi/mimo-v2.5-pro`, `deepseek/deepseek-v4-flash`). Subagent models
  are validated against the canonical catalog, which keys on the bare ids
  (`mimo-v2.5-pro`, `deepseek-v4-flash`) — matching hakaree's convention. Fixed
  both the runtime and subagent model ids.
- ci (agentcatalog): the global-manifest tests still looked up agents by the
  old `hakaree`/`kara` slugs; the rename changed them to
  `hakaree-software-engineer` / `kara-ui-and-graphics-designer`. This failure
  was masked in CI because validateManifests errored on the model first.
- lint (Go file length): driver.go (323) and handlers_alias_test.go (368)
  exceeded the 300-line ceiling. Split lifecycle ops into
  driver_lifecycle.go and the route-sync tests into handlers_alias_sync_test.go.
