+++ internal/agentcatalog/global_agents_test.go
@@ -23,7 +23,7 @@ func TestGlobalKaraManifestRequiresDesignPlugin(t *testing.T) {
 
 	var kara *Manifest
 	for index := range manifests {
-		if manifests[index].Slug == "kara" {
+		if manifests[index].Slug == "kara-ui-and-graphics-designer" {
 			kara = &manifests[index]
 			break
 		}
